Remove execute permission from xend-debug.log
The file /var/log/xen/xend-debug.log is currently being created with
executable permission bits set. This is because the os.open() method
defaults to using a mode of 0777 if no third parameter is provided.
The attached patch changes the mode to 0600 to ensure that the file
permissions come out as -rw-------  instead of -rwxr-xr-x

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>