xend: Don't use -q flag to bootloader if activating text console.
In Xen 3.0.3 series release, if one did 'xm create -c someguest' then
you would be shown the pygrub boot menu, allowing one to pick a
kernel. If you didn't use the -c arg, then it'd just pick the default
kernel.

In Xen 3.0.5 testing, xm is forcing the addition of the '-q' flag
whenever the 'bootargs' parameter is not given in the config file. So
if one does not have 'bootargs' specified, and one uses 'xm create -c
someguest' you never get to see the pygrub boot menu.

This patch fixes 'xm create' so that it does not add the '-q' flag
to the bootloader args, if the '-c' flag is given to xm create. This
makes the default user experiance operate closer to user
expectations.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>