[IA64] Backup/restore ACPI tables
We modify some of the ACPI tables for dom0 (limiting available CPUs,
modifying id/eid, and hiding SLIT/SRAT tables).  This causes problems
when we try to kexec with different dom0 CPU counts or from Xen to
Linux.  This introduces a mechanism to save ACPI tables before
modification and restoring them before kexec.

Signed-off-by: Alex Williamson <alex.williamson@hp.com>
Acked-by: Simon Horman <horms@verge.net.au>