[IA64] Automatic xenheap sizing
Automatically bump up the size of xenheap to account for the range
used by the boot allocator.  On systems like the HP superdome, when
cell local memory is enabled, the physical address space used for
memory gets much larger, causing the alloc_bitmap to slightly exceed
the default 64MB xenheap.  This is easily detected and prevented.

Signed-off-by: Alex Williamson <alex.williamson@hp.com>