ioemu: actually check for errors in bdrv_flush et al
bdrv_flush is declared to return void, but this is wrong because it
means that the implementations have nowhere to report their errors.
Indeed, the implementations generally ignore errors.

This patch corrects this by making it return int (implicitly, either 0
or -errno, as for other similar functions).  All of the
implementations and callers are adjusted too.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>